requirements.txt
google-adk
scikit-learn>=1.8.0
python-dotenv